+++ database/migrations/2025_02_13_073702_create_parts_table.php
@@ -16,7 +16,7 @@ return new class extends Migration {
         Schema::create('parts', function (Blueprint $table) {
             $table->uuid()->primary();
             $table->string('name');
-            $table->integer('qty');
+            $table->decimal('qty', 20, 2);
             $table->string('no_drawing')->nullable();
             $table->text('note')->nullable();
             $table->foreignIdFor(GlobalUnit::class);
