+++ database/migrations/2025_08_08_145937_add_column_to_const_mats_table.php
@@ -4,15 +4,14 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
         Schema::table('cons_mat_stds', function (Blueprint $table) {
-            $table->integer('qty')->default(0);
+            $table->decimal('qty', 20, 2)->default(0);
         });
     }
